loading the dataset, 1-hot vectors and categorical conversions
1. Load the dataset in Python.
(a) for the Penguin dataset, the features island and sex are strings, therefore they need to be converted
to numerical format to be fed to the MLP. To do so, experiment with 2 methods:
i. convert these features into 1-hot vectors (also known as dummy-coded data)
ii. convert these features into categories yourself